Run test, add visualization

+++ gss_filter.h
@@ -6,7 +6,7 @@
 
 float 	**	gauss_kernel(int map_size, float std_dev, int x_node, int y_node);
 
-void 		print_kernel(double **kernel, int map_size);
+void 		print_kernel(float **kernel, int map_size);
 
 
 #endif //__GSS_FILTER_H__
